+++ CHANGES.md
@@ -2,11 +2,13 @@
 
 ## [Unreleased]
 ### Added
+- `SwtExec` can now plays nicely with Kotlin coroutines.
 - `CoatMux.Layer` now exposes its root `Control` instance ([#12](https://github.com/diffplug/durian-swt/issues/12)).
 - `VScrollCtl` and `VScrollBubble` for easier VScroll handling ([#10](https://github.com/diffplug/durian-swt/issues/10)).
 ### Removed
 - **BREAKING** Removed `ColorPool.getSystemColor()` ([#8](https://github.com/diffplug/durian-swt/issues/8)).
 - **BREAKING** Switched type arguments of `TypedDataField` ([#3](https://github.com/diffplug/durian-swt/issues/3)).
+- **BREAKING** Removed OSGi metadata.
 
 ## [3.6.1] - 2022-02-18
 ### Fixed